% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{arsearch}
\alias{arsearch}
\title{Search for rigth seeds for the rigth AR simulation with arima.sin() finction using auto.arima() function}
\usage{
arsearch(
  a,
  z,
  n,
  ar11,
  ar22,
  ar33,
  p,
  d,
  q,
  sd = 1,
  j1,
  j2,
  j3,
  arr1,
  arr2,
  arr3
)
}
\arguments{
\item{a}{first seed boundary}

\item{z}{last seed boundary}

\item{n}{number of samples}

\item{ar11}{first coefficient of autoregressive}

\item{ar22}{second coefficient of autoregressive}

\item{ar33}{third coefficient of autoregressive}

\item{p}{order of the autoregressive}

\item{d}{degree of difference}

\item{q}{degree of moving average}

\item{sd}{standard deviation of the series}

\item{j1}{length of character to search for in first coefficient of autoregressive}

\item{j2}{length of character to search for in second coefficient of autoregressive}

\item{j3}{length of character to search for in second coefficient of autoregressive}

\item{arr1}{character to search for in first coefficient of autoregressive}

\item{arr2}{character to search for in second coefficient of autoregressive}

\item{arr3}{character to search for in third coefficient of autoregressive}
}
\description{
More detailed Description
}
\section{Functions}{
\itemize{
\item \code{arsearch}: ARSearch helps to Search for rigth seeds for the rigth AR simulation with arima.sin() finction using auto.arima() function
}}

